Add part of metadata editing functionality
App now can edit metadata for a single song. This is done by selecting
a song and tapping on the edit button that appears on the bottom right
of the screen.

ISSUES:
Although changes made to the metadata of a song are seen in the MP3
file, the changes are not seen in the media store (Android's data base
of media metadata). The fields for which the changes aren't reflected in
the media store are:
- Album art
- Genre
- Song recording year